Class "MasterPeptides"

A class to store the results of makeMaster. This class stored the 2 versions (orders) of the master final peptide data.

Objects from the Class

Objects can be created by calls of the form makeMaster.

Slots

masters:

Object of class "list" storing the 2 master data.frame objects.

pepfiles:

Object of class "character" with the list of final peptide input files.

fdr:

Object of class "numeric" with the peptide false discovery applied when creating the filter.

method:

Object of class "character" with the peptide p-value adjustment method. One of BH (default), qval or Bonferroni.

orders:

Object of class "list" with the numeric vectors specifying the order of pepfiles used to generate the respective masters data.frames.

fragmentfiles:

Object of class "character" with the list of final fragment input files.

fragments:

Object of class data.frame storing the combined final fragment data.

fragmentlibrary:

Object of class "MSnExp" storing the fragment library.

Methods

show

signature(object = "MasterPeptides"): to print a textual representation of the instance.

References

Improving qualitative and quantitative performance for MSE-based label free proteomics, N.J. Bond, P.V. Shliaha, K.S. Lilley and L. Gatto, Journal of Proteome Research, 2013, in press.

The Effects of Travelling Wave Ion Mobility Separation on Data Independent Acquisition in Proteomics Studies, P.V. Shliaha, N.J. Bond, L. Gatto and K.S. Lilley, Journal of Proteome Research, 2013, in press.