Computes FDR for all possible final peptide combinations

This function takes all possible combination of pepfiles of length greater or equal than 2 and computes the number of estimated incorrect peptides, the number of unique peptides, the number of unique protetypic peptides and the false discovery rate after merging for each combination. The best combination has an fdr lower than masterFdr and the highest number of unique (proteotypic) peptides.

estimateMasterFdr(pepfiles, fastafile, masterFdr = 0.025, fdr = 0.01,
  proteotypic = TRUE, missedCleavages = 0, IisL = FALSE,
  maxFileComb = length(pepfiles), verbose = interactive())

Arguments

pepfiles
A list of vector of final peptide filenames.

fastafile
A character with the fasta filename.

masterFdr
A numeric indicating the maximum merged false discovery to be allowed.

fdr
Peptide FDR level for individual peptide files filtering.

proteotypic
Logical. Should number proteotypic peptides be used to choose best combination and plot results or total number of unique peptides.

missedCleavages
Number of maximal missed cleavage sites. Default is 0.

IisL
If TRUE Isoleucin and Leucin are treated as identical. In this case sequences like "ABCI", "ABCL" are removed because they are not unqiue. If FALSE (default) "ABCI" and "ABCL" are reported as unique.

maxFileComb
A numeric to limit the accepted file combinations to reduce computation time. Default is length(pepfiles) meaning no limit set.

verbose
Should progress messages be printed?

Value

An instance of class "MasterFdrResults". See details above.

Details

The false discovery rate for the master (merged) file is calcualted by summing the number of estimated false discoveries for each individual final peptide file (number of unique peptides in that file multiplied by fdr) divided by the total number of unique peptides for that specific combination.

The function returns an instance of the class "MasterFdrResults".

References

Bond N. J., Shliaha P.V., Lilley K.S. and Gatto L., (2013) J. Prot. Research.

See also