Inspect peptide scores.

This function takes a final peptide file and returns information about the unique peptide scores and their number in each peptide matchType (PepFrag1 and PepFrag2) by protein dataBaseType (Random and Regular) category. The function is a lightweight verion of getPepNumbers and plotPepScores (to be used with Synapter instances) for individual files.

inspectPeptideScores(filename)

Arguments

filename
The name of a final peptide file.

Value

A table of peptide counts in each peptide matchType * protein dataBasteType category. Also plots the distribution of respective peptide scores.