Plot fragment matching results.

This method plots the results of the fragment matching procedure (fragmentMatching). A single plot contains two panels. The upper panel shows the identification fragments and the lower one the MS2 spectrum of the quantitation run. Common peaks are drawn in a slightly darker colour and with a point on the top.

Arguments

object
Object of class "Synapter" .

key
character, value to look for.

column
character, name of the column in which plotFragmentMatching looks for key.

…
Further arguments passed to internal functions.

Methods

signature(object = "Synapter", key = "character", column = "character", verbose = "logical", …)

Plots identification fragments against quantitation spectra. The … arguments are passed to the internal functions. Currently legend.cex, fragment.cex, and most of the plot.default arguments (like xlim, ylim, main, xlab, ylab, …) are supported. legend.cex and fragment.cex control the size of the legend and fragments labels (default: 0.5). Please see par for details about cex. If verbose = TRUE a progress bar is shown.

See also

fragmentMatching