Rescale for TOP3

This method rescales the intensity values of an MSnSet object to be suiteable for TOP3 quantification.

# S4 method for MSnSet,MSnSet
rescaleForTop3(before, after, saturationThreshold,
onlyForSaturatedRuns=TRUE, …)

Arguments

before
An MSnSet object before requantification.

after
The same MSnSet object as before but after requantification.

saturationThreshold
double, intensity of an ion (isotope of a given charge state) at which saturation is starting to occur.

onlyForSaturatedRuns
logical, rescale just runs where at least one isotope is affected by saturation.

…
further arguments passed to internal functions. Currently ignored.

Value

MSnSet where the assayData are requantified.

Details

If an MSnSet object was requantified using the method="sum" requantification method (see requantify,MSnSet-method) TOP3 is not valid anymore because the most abundant proteins are penalised by removing high intensity isotopes.

To overcome this rescaleForTop3 takes the proportion isotope/sum(isotopes for each requantified peptide and calculates a correction factor by comparing these proportions against the unsaturated isotopes before requantification. The new rescale intensity values of the isotopes are the mean correction factor multiplied with the corrected intensity values (see https://github.com/lgatto/synapter/issues/39#issuecomment-207987278 for the complete explanation/discussion).

References

See discussion on github: https://github.com/lgatto/synapter/issues/39

See also

MSnSet documentation: MSnSet https://github.com/lgatto/synapter/issues/39#issuecomment-207987278