Synapter/PLGS Agreement

This method checks the agreement between synapter analysis and PLGS results.

# S4 method for MSnSet
synapterPlgsAgreement(object, …)

Arguments

object
An MSnSet object.

…
further arguments, not used yet.

Value

MSnSet where the columns nIdentified, nAgree, nDisagree and synapterPlgsAgreementRatio were added to the feature data.

Details

Each synapter object has synapterPlgsAgreement column in its MatchedEMRTs data.frame (see writeMatchedEMRTs). After converting the synapter object into an MSnSet instance via as(synapterobject, "MSnSet" this column could be find in the feature data (fData(msnset)$synapterPlgsAgreement). In the synapterPlgsAgreement each peptide is classified as:

"agree": EMRT identified in identification and quantitation run by PLGS and same EMRT matched in synapter's grid search.

"disagree": EMRT identified in identification and quantitation run by PLGS and a different EMRT was matched in synapter's grid search.

"no_plgs_id": EMRT was not identified in the quantitation run by PLGS but matched in synapter's grid search.

"no_synapter_transfer": EMRT was identified in the identification and quantitation run by PLGS but not matched in synapter's grid search.

"no_id_or_transfer": EMRT was not identified in the quantitation run by PLGS and not matched in synapter's grid search.

"multiple_ident_matches": a single quantitation EMRT was matched by synapter to multiple identification EMRTs found by PLGS (could happen if the grid search parameters are too relaxed).

After combining multiple MSnSet the method synapterPlgsAgreement adds additional columns to the feature data:

nIdentified: how often a peptide was identified across multiple runs?

nAgree: how often a peptide was identified by PLGS and synapter across multiple runs (counts "agree" entries)?

nDisagree: how often a peptide was differently identified by PLGS and synapter across multiple runs (counts "disagree" entries)?

synapterPlgsAgreementRatio: nAgree/(nAgree + nDisagree).

References

See discussion on github: https://github.com/lgatto/synapter/issues/73

See also

MSnSet documentation: MSnSet