Loads a small test data for the 'synapter' package

Instead of using data to load the synapterTiny data set, synapterTinyData will load it and initialise it for proper downstream analysis, during which the 04_test_database.fasta file, provided with the package and references inside the object needs, to be accessed. However, as the exact location can not be known in advance, the reference is updated with the file's correct local path. This data set has been generated with the Synapter constructor. Note that the input data file sizes have been reduced by depleting many rows (peptides and EMRTs) from the original csv files. In addition, several columns that where not necessary for processing were also removed. As such, the data stored in synapterTiny does not relect the data obtained when following the section 'Preparing the input data' in the section, without however affecting the processing and final results.

synapterTinyData()

Value

A character vector with the data set name, "synapterTiny". Used for its side effect of loading synapterTiny, an instance of class Synapter, in .GlovalEnv.

Source

Improving qualitative and quantitative performance for MSE-based label free proteomics, N.J. Bond, P.V. Shliaha, K.S. Lilley and L. Gatto, Journal of Proteome Research, 2013, in press. The Effects of Travelling Wave Ion Mobility Separation on Data Independent Acquisition in Proteomics Studies, P.V. Shliaha, N.J. Bond, L. Gatto and K.S. Lilley, Journal of Proteome Research, 2013, in press.

Examples

synapterTinyData()
synapterTiny

#> Object of class "Synapter" 
#> Class version 2.0.0 
#> Package version 1.99.0 
#> Data files:
#>  + Identification pep file: 01_HDMSe_tiny.csv 
#>  + Quantitation pep file: 02_MSe_tiny.csv 
#>  + Quantitation Pep3DAMRT file: 03_Pep3D_tiny.csv 
#>  + Fasta file: 04_test_database.fasta 
#> Log:
#> [1] "Instance created on Wed Jun  6 23:37:36 2012"
#> [2] "Read identification peptide data [5915,18]"  
#> [ 9 lines ]
#> [12] "Filtered identification Random entries [4979,22]"
#> [13] "Instance updated to synapter 1.99.0 on Sun Oct 16 16:47:24 2016"