Intensity requantification

This method tries to remove saturation effects from the intensity counts.

# S4 method for MSnSet
requantify(object, saturationThreshold,
method=c("sum", "reference", "th.mean", "th.median",
"th.weighted.mean"), …)

Arguments

object
An MSnSet object.

saturationThreshold
double, intensity of an ion (isotope of a given charge state) at which saturation is starting to occur.

method
character, requantification method, please see details section.

…
further arguments passed to internal functions. Currently onlyCommonIsotopes for method="sum" and requantifyAll for method=c("th.mean", "th.median", "th.weighted.mean") are supported.

Value

MSnSet where the assayData are requantified.

Details

Currently requantify supports 3 (5) different requantification methods.

"sum" is the simplest requantification method. All ions of a peptide below the saturation threshold are summed to get the new intensity. This method accept an additional argument, namely onlyCommonIsotopes If onlyCommonIsotopes=TRUE (default) all ions that are not seen in all runs are removed and only the common seen ions are summed. In contrast onlyCommonIsotopes=FALSE sums all ions regardless they are present in all runs.

In "reference" the run that has the most unsaturated ions in common with all the other runs. If there are more than one run, the most intense is used as reference. The other runs are corrected as follows:

Find common ions between current and the reference run.

Divide the intensities of the common ions and calculate the mean of these quotients as a run specific scaling factor.

Multiply the unsaturated ions of the current run by the scaling factor and replace the saturated ones by the product of the scaling factor and the intensities of their corresponding ions in the reference run.

Sum the rescaled ion intensities.

The "th.*" methods are nearly identical. All of them calculate the theoretical isotopic distribution for the given sequence of the peptide. Subsequently the unsaturated ions are divided by their theoretical proportion and the mean/median/weighted.mean (proportions are used as weights) of these intensities are calculated per charge state. The sum of the charge state values is used as requantified intensity for this peptide.
If requantifyAll=FALSE (default) just peptides with at least one saturated ion are requantified (unsaturated peptides are unaffected). If requantify=TRUE all peptides even these where all ions are below saturationThreshold are requantified by their theoretical distribution.

References

See discussion on github: https://github.com/lgatto/synapter/issues/39

See also