Synergise identification and quantitation results

Performs a complete default analysis on the files defined in filenames, creates a complete html report and saves/exports all results as csv and rds files. See details for a description of the pipeline and Synapter for manual execution of individual steps.

synergise(filenames, master = FALSE, object, outputdir,
  outputfile = paste0("synapter_report_", strftime(Sys.time(),
  "%Y%m%d-%H%M%S"), ".html"), fdr = 0.01, fdrMethod = c("BH",
  "Bonferroni", "qval"), fpr = 0.01, peplen = 7, missedCleavages = 0,
  IisL = FALSE, identppm = 20, quantppm = 20, uniquepep = TRUE,
  span = 0.05, grid.ppm.from = 2, grid.ppm.to = 20, grid.ppm.by = 2,
  grid.nsd.from = 0.5, grid.nsd.to = 5, grid.nsd.by = 0.5,
  grid.subset = 1, grid.n = 0, grid.param.sel = c("auto", "model",
  "total", "details"), mergedEMRTs = c("rescue", "copy", "transfer"),
  template = system.file("reports", "synergise1.Rmd", package = "synapter"),
  verbose = interactive())

Arguments

filenames
A named list of file names to be load. The names must be identpeptide, quantpeptide, quantpep3d and fasta (could be an RDS file created by link{createUniquePeptideDbRds}). identpeptide can be a csv final peptide file (from PLGS) or a saved "MasterPeptides" data object as created by makeMaster if working with master peptide data. To serialise the "MasterPeptides" instance, use the saveRDS function, and file extenstion rds.

master
A logical indicating if the identification final peptide files are master (see makeMaster) or regular files. Default is FALSE.

object
An instance of class Synapter that will be copied, processed and returned. If filenames are also provided, the latter and object's inputFiles will be checked for equality.

outputdir
A character with the full path to an existing directory.

outputfile
A character with the file name for the report.

fdr
Peptide false discovery rate. Default is 0.01.

fdrMethod
P-value adjustment method. One of "BH" (default) for Benjamini and HochBerg (1995), "Bonferroni" for Bonferroni's single-step adjusted p-values for strong control of the FWER and "qval" from the qvalue package. See Synapter for references.

fpr
Protein false positive rate. Default is 0.01.

peplen
Minimum peptide length. Default is 7.

missedCleavages
Number of allowed missed cleavages. Default is 0.

IisL
If TRUE Isoleucin and Leucin are treated as equal. In this case sequences like "ABCI", "ABCL" are removed because they are not unqiue. If FALSE (default) "ABCI" and "ABCL" are reported as unique.

identppm
Identification mass tolerance (in ppm). Default is 20.

quantppm
Quantitation mass tolerance (in ppm). Default is 20.

uniquepep
A logical is length 1 indicating if only unique peptides in the identification and quantitation peptides as well as unique tryptic peptides as defined in the fasta file. Default is TRUE.

span
The loess span parameter. Default is 0.05.

grid.ppm.from
Mass tolerance (ppm) grid starting value. Default is 2.

grid.ppm.to
Mass tolerance (ppm) grid ending value. Default is 20.

grid.ppm.by
Mass tolerance (ppm) grid step value. Default is 2.

grid.nsd.from
Number of retention time stdev grid starting value. Default is 0.5.

grid.nsd.to
Number of retention time stdev grid ending value. Default is 5.

grid.nsd.by
Number of retention time stdev grid step value. Default is 0.5.

grid.subset
Percentage of features to be used for the grid search. Default is 1.

grid.n
Absolute number of features to be used for the grid search. Default is 0, i.e ignored.

grid.param.sel
Grid parameter selection method. One of auto (default), details, model or total. See Synapter for details on these selection methods.

mergedEMRTs
One of "rescue" (default), "copy" or "transfer". See the documentation for the findEMRTs function in Synapter for details.

template
A character full path to Rmd template.

verbose
A logical indicating if progress output should be printed to the console. Default is TRUE.

Value

Invisibly returns an object of class Synapter. Used for its side effect of creating an html report of the run in outputdir.

Details

Data can be input as a Synapter object if available or as a list of files (see filenames) that will be used to read the data in. The html report and result files will be created in the outputdir folder. All other input parameters have default values.

The data processing and analysis pipeline is as follows:

If uniquepep is set to TRUE (default), only unique proteotypic identification and quantitation peptides are retained.

Peptides are filtered for a FDR <= fdr (default is 0.01) using the "BH" method (see fdr and fdrMethod parameters for details).

Peptide with a mass tolerance > 20 ppms (see quantppm and identppm) are filtered out.

Peptides with a protein false positive rate (as reported by the PLGS software) > fpr are filtered out.

Common identification and quantitation peptides are merged and a retention time model is created using the Local Polynomial Regression Fitting (loess function for the stats package) using a default span value of 0.05.

A grid search to optimise the width in retention time and mass tolerance for EMRTs matching is performed. The default grid search space is from 0.5 to 5 by 0.5 retention time model standard deviations (see grid.nsd.from, grid.nsd.to and grid.nsd.by parameters) and from 2 to 20 by 2 parts per million (ppm) for mass tolerance (see grid.ppm.from, grid.ppm.to and grid.ppm.by parameters). The data can be subset using using an absolute number of features (see grid.n) or a fixed percentage (see grid.subset). The pair of optimal nsd and ppm is chosen (see grid.param.sel parameter).

The quantitation EMRTs are matched using the optimised parameters.

If a master identification file is used (master is set to TRUE, default is FALSE), the relevant actions that have already been executed when the file was created with makeMaster are not repeated here.

References

Bond N. J., Shliaha P.V., Lilley K.S. and Gatto L. (2013) J. Prot. Research.

Examples

output <- tempdir() ## a temporary directory
synapterTinyData()
synergise(object = synapterTiny, outputdir = output, outputfile = "synapter.html",
          grid.subset = 0.2)

#> 
#> 
#> processing file: synergise1.Rmd

#>   |                                                                         |                                                                 |   0%  |                                                                         |.                                                                |   2%
#>    inline R code fragments
#> 
#>   |                                                                         |..                                                               |   3%
#> label: setup (with options) 
#> List of 1
#>  $ echo: logi FALSE
#> 
#>   |                                                                         |...                                                              |   5%
#>   ordinary text without R code
#> 
#>   |                                                                         |....                                                             |   6%
#> label: input.files (with options) 
#> List of 2
#>  $ echo   : logi FALSE
#>  $ results: chr "asis"
#> 
#>   |                                                                         |.....                                                            |   8%
#>   ordinary text without R code
#> 
#>   |                                                                         |......                                                           |   9%
#> label: input.parameters (with options) 
#> List of 2
#>  $ echo   : logi FALSE
#>  $ results: chr "asis"
#> 
#>   |                                                                         |.......                                                          |  11%
#>   ordinary text without R code
#> 
#>   |                                                                         |........                                                         |  12%
#> label: input.gridparameters (with options) 
#> List of 2
#>  $ echo   : logi FALSE
#>  $ results: chr "asis"
#> 
#>   |                                                                         |.........                                                        |  14%
#>   ordinary text without R code
#> 
#>   |                                                                         |..........                                                       |  15%
#> label: synergise.filtering.pepScores (with options) 
#> List of 1
#>  $ echo: logi FALSE
#> 

#>   |                                                                         |...........                                                      |  17%
#>   ordinary text without R code
#> 
#>   |                                                                         |............                                                     |  18%
#> label: synergise.filtering.pepNumbers (with options) 
#> List of 1
#>  $ echo: logi FALSE
#> 
#>   |                                                                         |.............                                                    |  20%
#>   ordinary text without R code
#> 
#>   |                                                                         |..............                                                   |  21%
#> label: synergise.filtering.fdr (with options) 
#> List of 1
#>  $ echo: logi FALSE
#> 

#>   |                                                                         |...............                                                  |  23%
#>   ordinary text without R code
#> 
#>   |                                                                         |................                                                 |  24%
#> label: synergise.filtering.data (with options) 
#> List of 1
#>  $ echo: logi FALSE
#> 
#>   |                                                                         |.................                                                |  26%
#>   ordinary text without R code
#> 
#>   |                                                                         |..................                                               |  27%
#> label: synergise.filtering.ppmErrorPre (with options) 
#> List of 1
#>  $ echo: logi FALSE
#> 

#>   |                                                                         |...................                                              |  29%
#>   ordinary text without R code
#> 
#>   |                                                                         |....................                                             |  30%
#> label: synergise.filtering.ppmErrorQsPre (with options) 
#> List of 2
#>  $ echo   : logi FALSE
#>  $ results: chr "asis"
#> 
#>   |                                                                         |.....................                                            |  32%
#>   ordinary text without R code
#> 
#>   |                                                                         |......................                                           |  33%
#> label: synergise.filtering.ppm (with options) 
#> List of 1
#>  $ echo: logi FALSE
#> 
#>   |                                                                         |.......................                                          |  35%
#>   ordinary text without R code
#> 
#>   |                                                                         |........................                                         |  36%
#> label: synergise.filtering.ppmErrorPost (with options) 
#> List of 1
#>  $ echo: logi FALSE
#> 

#>   |                                                                         |.........................                                        |  38%
#>   ordinary text without R code
#> 
#>   |                                                                         |..........................                                       |  39%
#> label: synergise.filtering.ppmErrorQsPost (with options) 
#> List of 2
#>  $ echo   : logi FALSE
#>  $ results: chr "asis"
#> 
#>   |                                                                         |...........................                                      |  41%
#>   ordinary text without R code
#> 
#>   |                                                                         |............................                                     |  42%
#> label: synergise.filtering.fpr (with options) 
#> List of 1
#>  $ echo: logi FALSE
#> 
#>   |                                                                         |.............................                                    |  44%
#>   ordinary text without R code
#> 
#>   |                                                                         |..............................                                   |  45%
#> label: synergise.merge (with options) 
#> List of 1
#>  $ echo: logi FALSE
#> 
#>   |                                                                         |...............................                                  |  47%
#>   ordinary text without R code
#> 
#>   |                                                                         |................................                                 |  48%
#> label: synergise.rtmodel.plotRt (with options) 
#> List of 1
#>  $ echo: logi FALSE
#> 

#>   |                                                                         |................................                                 |  50%
#>   ordinary text without R code
#> 
#>   |                                                                         |.................................                                |  52%
#> label: synergise.rtmodel.modelling (with options) 
#> List of 1
#>  $ echo: logi FALSE
#> 
#>   |                                                                         |..................................                               |  53%
#>   ordinary text without R code
#> 
#>   |                                                                         |...................................                              |  55%
#> label: synergise.rtmodel.rtdiffs (with options) 
#> List of 1
#>  $ echo: logi FALSE
#> 

#>   |                                                                         |....................................                             |  56%
#>   ordinary text without R code
#> 
#>   |                                                                         |.....................................                            |  58%
#> label: synergise.rtmodel.rtdiffsTable (with options) 
#> List of 2
#>  $ echo   : logi FALSE
#>  $ results: chr "asis"
#> 
#>   |                                                                         |......................................                           |  59%
#>   ordinary text without R code
#> 
#>   |                                                                         |.......................................                          |  61%
#> label: rtmodel.plotAllFeatures (with options) 
#> List of 1
#>  $ echo: logi FALSE
#> 

#>   |                                                                         |........................................                         |  62%
#>   ordinary text without R code
#> 
#>   |                                                                         |.........................................                        |  64%
#> label: synergise.rtmodel.plotSomeFeatures (with options) 
#> List of 3
#>  $ echo   : logi FALSE
#>  $ dev    : chr "svg"
#>  $ fig.cap: chr "Using nsd = 2 and ppm error = 10 as an example.<br />Optimal values are selected below."
#> 

#>   |                                                                         |..........................................                       |  65%
#>    inline R code fragments
#> 
#>   |                                                                         |...........................................                      |  67%
#> label: grid.comment (with options) 
#> List of 3
#>  $ echo   : logi FALSE
#>  $ eval   : logi FALSE
#>  $ results: chr "hide"
#> 
#>   |                                                                         |............................................                     |  68%
#>    inline R code fragments
#> 

#> No ion mobility available. Step back to 2D grid search.

#>   |                                                                         |                                                                 |   0%  |                                                                         |=                                                                |   1%  |                                                                         |=                                                                |   2%  |                                                                         |==                                                               |   3%  |                                                                         |===                                                              |   4%  |                                                                         |===                                                              |   5%  |                                                                         |====                                                             |   6%  |                                                                         |=====                                                            |   7%  |                                                                         |=====                                                            |   8%  |                                                                         |======                                                           |   9%  |                                                                         |======                                                           |  10%  |                                                                         |=======                                                          |  11%  |                                                                         |========                                                         |  12%  |                                                                         |========                                                         |  13%  |                                                                         |=========                                                        |  14%  |                                                                         |==========                                                       |  15%  |                                                                         |==========                                                       |  16%  |                                                                         |===========                                                      |  17%  |                                                                         |============                                                     |  18%  |                                                                         |============                                                     |  19%  |                                                                         |=============                                                    |  20%  |                                                                         |==============                                                   |  21%  |                                                                         |==============                                                   |  22%  |                                                                         |===============                                                  |  23%  |                                                                         |================                                                 |  24%  |                                                                         |================                                                 |  25%  |                                                                         |=================                                                |  26%  |                                                                         |==================                                               |  27%  |                                                                         |==================                                               |  28%  |                                                                         |===================                                              |  29%  |                                                                         |====================                                             |  30%  |                                                                         |====================                                             |  31%  |                                                                         |=====================                                            |  32%  |                                                                         |=====================                                            |  33%  |                                                                         |======================                                           |  34%  |                                                                         |=======================                                          |  35%  |                                                                         |=======================                                          |  36%  |                                                                         |========================                                         |  37%  |                                                                         |=========================                                        |  38%  |                                                                         |=========================                                        |  39%  |                                                                         |==========================                                       |  40%  |                                                                         |===========================                                      |  41%  |                                                                         |===========================                                      |  42%  |                                                                         |============================                                     |  43%  |                                                                         |=============================                                    |  44%  |                                                                         |=============================                                    |  45%  |                                                                         |==============================                                   |  46%  |                                                                         |===============================                                  |  47%  |                                                                         |===============================                                  |  48%  |                                                                         |================================                                 |  49%  |                                                                         |================================                                 |  50%  |                                                                         |=================================                                |  51%  |                                                                         |==================================                               |  52%  |                                                                         |==================================                               |  53%  |                                                                         |===================================                              |  54%  |                                                                         |====================================                             |  55%  |                                                                         |====================================                             |  56%  |                                                                         |=====================================                            |  57%  |                                                                         |======================================                           |  58%  |                                                                         |======================================                           |  59%  |                                                                         |=======================================                          |  60%  |                                                                         |========================================                         |  61%  |                                                                         |========================================                         |  62%  |                                                                         |=========================================                        |  63%  |                                                                         |==========================================                       |  64%  |                                                                         |==========================================                       |  65%  |                                                                         |===========================================                      |  66%  |                                                                         |============================================                     |  67%  |                                                                         |============================================                     |  68%  |                                                                         |=============================================                    |  69%  |                                                                         |==============================================                   |  70%  |                                                                         |==============================================                   |  71%  |                                                                         |===============================================                  |  72%  |                                                                         |===============================================                  |  73%  |                                                                         |================================================                 |  74%  |                                                                         |=================================================                |  75%  |                                                                         |=================================================                |  76%  |                                                                         |==================================================               |  77%  |                                                                         |===================================================              |  78%  |                                                                         |===================================================              |  79%  |                                                                         |====================================================             |  80%  |                                                                         |=====================================================            |  81%  |                                                                         |=====================================================            |  82%  |                                                                         |======================================================           |  83%  |                                                                         |=======================================================          |  84%  |                                                                         |=======================================================          |  85%  |                                                                         |========================================================         |  86%  |                                                                         |=========================================================        |  87%  |                                                                         |=========================================================        |  88%  |                                                                         |==========================================================       |  89%  |                                                                         |==========================================================       |  90%  |                                                                         |===========================================================      |  91%  |                                                                         |============================================================     |  92%  |                                                                         |============================================================     |  93%  |                                                                         |=============================================================    |  94%  |                                                                         |==============================================================   |  95%  |                                                                         |==============================================================   |  96%  |                                                                         |===============================================================  |  97%  |                                                                         |================================================================ |  98%  |                                                                         |================================================================ |  99%  |                                                                         |=================================================================| 100%
#>   |                                                                         |.............................................                    |  70%
#> label: grid.table (with options) 
#> List of 2
#>  $ echo   : logi FALSE
#>  $ results: chr "asis"
#> 

#>   |                                                                         |..............................................                   |  71%
#>   ordinary text without R code
#> 
#>   |                                                                         |...............................................                  |  73%
#> label: grid.details (with options) 
#> List of 2
#>  $ echo   : logi FALSE
#>  $ results: chr "asis"
#> 
#>   |                                                                         |................................................                 |  74%
#>   ordinary text without R code
#> 
#>   |                                                                         |.................................................                |  76%
#> label: synergise.grid.bestParamaters (with options) 
#> List of 2
#>  $ echo   : logi FALSE
#>  $ results: chr "asis"
#> 
#>   |                                                                         |..................................................               |  77%
#>    inline R code fragments
#> 
#>   |                                                                         |...................................................              |  79%
#> label: synergise.emrt.find (with options) 
#> List of 1
#>  $ echo: logi FALSE
#> 

#>   |                                                                         |....................................................             |  80%
#>   ordinary text without R code
#> 
#>   |                                                                         |.....................................................            |  82%
#> label: synergise.emrt.table (with options) 
#> List of 2
#>  $ echo   : logi FALSE
#>  $ results: chr "asis"
#> 
#>   |                                                                         |......................................................           |  83%
#>   ordinary text without R code
#> 
#>   |                                                                         |.......................................................          |  85%
#> label: synergise.performance.enrichment (with options) 
#> List of 2
#>  $ echo   : logi FALSE
#>  $ results: chr "asis"
#> 
#>   |                                                                         |........................................................         |  86%
#>   ordinary text without R code
#> 
#>   |                                                                         |.........................................................        |  88%
#> label: synergise.performance.overlap (with options) 
#> List of 2
#>  $ echo   : logi FALSE
#>  $ results: chr "asis"
#> 
#>   |                                                                         |..........................................................       |  89%
#>   ordinary text without R code
#> 
#>   |                                                                         |...........................................................      |  91%
#> label: synergise.performance.source (with options) 
#> List of 2
#>  $ echo   : logi FALSE
#>  $ results: chr "asis"
#> 
#>   |                                                                         |............................................................     |  92%
#>   ordinary text without R code
#> 
#>   |                                                                         |.............................................................    |  94%
#> label: synergise.export (with options) 
#> List of 1
#>  $ echo: logi FALSE
#> 
#>   |                                                                         |..............................................................   |  95%
#>   ordinary text without R code
#> 
#>   |                                                                         |...............................................................  |  97%
#> label: synergise.log (with options) 
#> List of 2
#>  $ echo   : logi FALSE
#>  $ results: chr "asis"
#> 
#>   |                                                                         |................................................................ |  98%
#>   ordinary text without R code
#> 
#>   |                                                                         |.................................................................| 100%
#> label: synergise.sessionInfo (with options) 
#> List of 1
#>  $ echo: logi FALSE
#> 
#> 

#> output file: synergise1.knit.md

#> /usr/bin/pandoc +RTS -K512m -RTS synergise1.utf8.md --to html --from markdown+autolink_bare_uris+ascii_identifiers+tex_math_single_backslash --output /tmp/RtmpaV2ulX/synapter.html --smart --email-obfuscation none --self-contained --standalone --section-divs --table-of-contents --toc-depth 3 --variable toc_float=1 --variable toc_selectors=h1,h2,h3 --variable toc_print=1 --template /home/sebastian/R/x86_64-pc-linux-gnu-library/3.3/rmarkdown/rmd/h/default.html --no-highlight --variable highlightjs=1 --variable 'theme:bootstrap' --include-in-header /tmp/RtmpaV2ulX/rmarkdown-str2ed833864d2b.html 

#> 
#> Output created: /tmp/RtmpaV2ulX/synapter.html

htmlReport <- paste0("file:///", file.path(output, "synapter.html")) ## the result report
## Not run: ------------------------------------
# browseURL(htmlReport) ## open the report with default browser
## ---------------------------------------------